Update vendored MNX to schema version 34 (bpm allows floats) (#64)
Refresh third_party/w3c-mnx via scripts/update-w3c-mnx.sh, moving the
upstream pin from e1570dd to 0a8c760. The schema $id advances from
version/33 to version/34.

The only substantive schema change is that `bpm` widens from "integer"
to "number":

    "bpm": { "exclusiveMinimum": 0, "type": "integer" -> "number" }

No example JSON changed (still 52 files), so the kSchemaOnlyExamples
exemptions are unaffected.

Accordingly, global::Tempo now models bpm as a double: the Required
field, the constructor, make(), and MNX_REQUIRED_PROPERTY. Integer-
encoded bpm values in existing documents continue to read back
correctly, since nlohmann converts on get<double>().

Adds Global.TempoFractionalBpm, which covers a fractional bpm
round-trip, schema validation of both fractional and whole-number
values, and backward compatibility with integer-encoded input. Tempo
previously had no test coverage.

+++ src/Global.h
@@ -281,7 +281,7 @@ public:
     /// @brief initializer class for #Tempo
     struct Required
     {
-        int bpm{};                       ///< the beats per minute
+        double bpm{};                    ///< the beats per minute
         NoteValue::Required noteValue{};   ///< the note value for the tempo
     };
 
@@ -296,7 +296,7 @@ public:
     /// @param key The JSON key to use for embedding in parent.
     /// @param bpm The beats per minute
     /// @param noteValue The note value for the tempo
-    Tempo(Base& parent, std::string_view key, int bpm, const NoteValue::Required& noteValue)
+    Tempo(Base& parent, std::string_view key, double bpm, const NoteValue::Required& noteValue)
         : ArrayElementObject(parent, key)
     {
         set_bpm(bpm);
@@ -307,9 +307,9 @@ public:
     operator Required() const { return { bpm(), value() }; }
 
     /// @brief Create a Required instance for #Tempo.
-    static Required make(int bpm, const NoteValue::Required& noteValue) { return { bpm, noteValue }; }
+    static Required make(double bpm, const NoteValue::Required& noteValue) { return { bpm, noteValue }; }
 
-    MNX_REQUIRED_PROPERTY(int, bpm);                ///< the beats per minute of this tempo marking
+    MNX_REQUIRED_PROPERTY(double, bpm);             ///< the beats per minute of this tempo marking
     MNX_OPTIONAL_CHILD(RhythmicPosition, location,
         (const FractionValue&, position)); ///< location within the measure of the tempo marking
     MNX_REQUIRED_CHILD(NoteValue, value,
